requirements.txt
appdirs==1.4.3
click==6.7
DateTime==4.1.1
Flask==0.12.1
happybase==1.1.0
itsdangerous==0.24
Jinja2==2.9.6
logging==0.4.9.6
MarkupSafe==1.0
MySQL-python==1.2.5
packaging==16.8
ply==3.10
pyparsing==2.2.0
python-dateutil==2.6.0
python-memcached==1.58
pytz==2017.2
PyYAML==3.12
six==1.10.0
thriftpy==0.3.9
timelib==0.2.4
uWSGI==2.0.15
var-dump==1.1
Werkzeug==0.12.1
zope.interface==4.3.3
